added second type of ml handler interface previous is supported also
