+++ mindsdb/integrations/handlers/byom_handler/__init__.py
@@ -1,6 +1,6 @@
 from mindsdb.integrations.libs.const import HANDLER_TYPE
 
-from .byom_handler_exec import BYOMHandler_EXECUTOR as Handler, connection_args
+from .byom_handler import BYOMHandler as Handler, connection_args
 from .__about__ import __version__ as version, __description__ as description
 
 title = 'BYOM'
