+++ mindsdb/integrations/libs/base.py
@@ -125,7 +125,7 @@ class BaseMLEngine:
     An instance of this class will be generated, used, and destroyed for each interaction with the underlying framework.
     """
 
-    def __init__(self, model_storage, metadata_storage) -> None:
+    def __init__(self, model_storage, engine_storage) -> None:
         """
         Initialize any objects, fields or parameters required by the ML engine.
 
@@ -134,7 +134,7 @@ class BaseMLEngine:
             - metadata_storage: stores all model metadata in an internal MindsDB database.
         """
         self.model_storage = model_storage
-        self.metadata_storage = metadata_storage
+        self.engine_storage = engine_storage
 
     def create_engine(self, connection_args: dict):
         """
